Try to fix typing problems between 32bit and 64bit platforms.

+++ src/content.c
@@ -110,8 +110,8 @@ void strip_html_markup(bstring str) {
 }
 
 // Dan Bernstein's djb2 hash function
-unsigned long long hash(char *str) {
-  long long hash = 5381;
+uint64_t hash(char *str) {
+  uint64_t hash = 5381;
 
   while(*str!='\0') {
     int c = *str;
